use new CRC16 implementation and smaller enhancements
The CRC16 check is implemented with the correct CRC16 algorithm, but
checked in two byte orders because some smart-meters violate the specs
and write in wrong order. Bother are checked and if one match it is
considered as valid.
Beside the CRC16 stuff there are some fixes regarding the extra added
filler bytes at the end of the message. and the CRC is no longer
written in „toString“ to be better able to compare the messages for
tests.