add more real-live sample data and enhance testing-flow
Testing flow now is a bit documented :-)
Additionally the re-written SML is parsed again and the string versions
of the first read and this re-read messages are compared too.
SO these sample data test the reading, the writing and parsing the
self-written version to make sure it matches with the original message

Additionally logic is implemented to ignore checksum-differences the
the SML messages while comparing.

Sample-Data #26 still fails because of the int64 topic … will be fixed
next.